알림탭 삭제 , 애플로그인 수정

+++ haruGomin/logIn/selectGomin/Network/signUpResponse.swift
@@ -7,7 +7,7 @@
 //
 
 struct Profile:Codable {
-    let userId:Int?
+    let userId:Int64?
     let nickname:String?
     let profileImage:String?
     let ageRange:Int?
